Show CD icon for cards without a back side
Previously a back side was required to make the icon show up. Now it is displayed on all .rem-container--concept-rem-type.

I tried to reduce duplication by moving the masking rules to a single block with `background: var(--no-practice-color)` as default color.
You may need to pick a good color here.

The concept icon is also used for plain cards. Do we want a different icon here?